Move transactions to a separate file
